Desktop/ourdatav5/requirements_finetuning.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.7.0
pefltudio>=0.0.1
tqdm
numpy
pandas
scikit-learn
openpyxl
pdfplumber
python-docx
